Introduce Procedure and Measure system
This is used mainly for implementing autonomous mode control. A
Procedure can have multiple Steps, which consist of a Coordinator and a
Measure which, when complete, will trigger a change to the next Step in
the Procedure, until the Procedure is out of Steps.

A Procedure also has a global Coordinator that is always applied.
Traditional modes still work, and are passed through as Procedures with
no Steps and the Coordinator as the global Coordinator of the Procedure.